refactor dwelling to move datatypes for xml extracted data to extracted_datatypes.py
