+++ python/src/energuide/dwelling.py
@@ -4,6 +4,7 @@ import typing
 from dateutil import parser
 import cerberus
 from energuide import reader
+from energuide.extracted_datatypes import Floor, Ceiling
 
 
 class NoInputDataException(Exception):
@@ -65,86 +66,6 @@ class Region(enum.Enum):
         return output
 
 
-class _Ceiling(typing.NamedTuple):
-    label: typing.Optional[str]
-    type_english: typing.Optional[str]
-    type_french: typing.Optional[str]
-    nominal_rsi: typing.Optional[float]
-    effective_rsi: typing.Optional[float]
-    area_metres: typing.Optional[float]
-    length_metres: typing.Optional[float]
-
-
-class _Floor(typing.NamedTuple):
-    label: typing.Optional[str]
-    nominal_rsi: typing.Optional[float]
-    effective_rsi: typing.Optional[float]
-    area_metres: typing.Optional[float]
-    length_metres: typing.Optional[float]
-
-
-_RSI_MULTIPLIER = 5.678263337
-_FEET_MULTIPLIER = 3.28084
-_FEET_SQUARED_MULTIPLIER = _FEET_MULTIPLIER**2
-
-
-class Ceiling(_Ceiling):
-    @classmethod
-    def from_data(cls, ceiling: typing.Dict[str, typing.Any]):
-        return Ceiling(
-            label=ceiling['label'],
-            type_english=ceiling['typeEnglish'],
-            type_french=ceiling['typeFrench'],
-            nominal_rsi=ceiling['nominalRsi'],
-            effective_rsi=ceiling['effectiveRsi'],
-            area_metres=ceiling['area'],
-            length_metres=ceiling['length']
-
-        )
-
-    def to_dict(self):
-        return {
-            'label': self.label,
-            'typeEnglish': self.type_english,
-            'typeFrench': self.type_french,
-            'nominalRsi': self.nominal_rsi,
-            'nominalR': (self.nominal_rsi * _RSI_MULTIPLIER) if self.nominal_rsi is not None else None,
-            'effectiveRsi': self.effective_rsi,
-            'effectiveR': (self.effective_rsi * _RSI_MULTIPLIER) if self.effective_rsi is not None else None,
-            'areaMetres': self.area_metres,
-            'areaFeet': (self.area_metres * _FEET_SQUARED_MULTIPLIER) if self.area_metres is not None else None,
-            'lengthMetres': self.length_metres,
-            'lengthFeet': (self.length_metres * _FEET_MULTIPLIER) if self.length_metres is not None else None
-        }
-
-
-class Floor(_Floor):
-
-    @classmethod
-    def from_data(cls, ceiling: typing.Dict[str, typing.Any]):
-        return Floor(
-            label=ceiling['label'],
-            nominal_rsi=ceiling['nominalRsi'],
-            effective_rsi=ceiling['effectiveRsi'],
-            area_metres=ceiling['area'],
-            length_metres=ceiling['length']
-
-        )
-
-    def to_dict(self):
-        return {
-            'label': self.label,
-            'nominalRsi': self.nominal_rsi,
-            'nominalR': (self.nominal_rsi * _RSI_MULTIPLIER) if self.nominal_rsi is not None else None,
-            'effectiveRsi': self.effective_rsi,
-            'effectiveR': (self.effective_rsi * _RSI_MULTIPLIER) if self.effective_rsi is not None else None,
-            'areaMetres': self.area_metres,
-            'areaFeet': (self.area_metres * _FEET_SQUARED_MULTIPLIER) if self.area_metres is not None else None,
-            'lengthMetres': self.length_metres,
-            'lengthFeet': (self.length_metres * _FEET_MULTIPLIER) if self.length_metres is not None else None
-        }
-
-
 class _ParsedDwellingDataRow(typing.NamedTuple):
     eval_id: int
     eval_type: EvaluationType
@@ -172,37 +93,8 @@ class ParsedDwellingDataRow(_ParsedDwellingDataRow):
         'forwardSortationArea': {'type': 'string', 'required': True, 'regex': '[A-Z][0-9][A-Z]'},
         'HOUSEREGION': {'type': 'string', 'required': True},
 
-        'ceilings': {
-            'type': 'list',
-            'required': True,
-            'schema': {
-                'type': 'dict',
-                'schema': {
-                    'label': {'type': 'string', 'required': True},
-                    'typeEnglish': {'type': 'string', 'required': True},
-                    'typeFrench': {'type': 'string', 'required': True},
-                    'nominalRsi': {'type': 'float', 'required': True, 'coerce': float},
-                    'effectiveRsi': {'type': 'float', 'required': True, 'coerce': float},
-                    'area': {'type': 'float', 'required': True, 'coerce': float},
-                    'length': {'type': 'float', 'required': True, 'coerce': float}
-                }
-            }
-        },
-
-        'floors': {
-            'type': 'list',
-            'required': True,
-            'schema': {
-                'type': 'dict',
-                'schema': {
-                    'label': {'type': 'string', 'required': True},
-                    'nominalRsi': {'type': 'float', 'required': True, 'coerce': float},
-                    'effectiveRsi': {'type': 'float', 'required': True, 'coerce': float},
-                    'area': {'type': 'float', 'required': True, 'coerce': float},
-                    'length': {'type': 'float', 'required': True, 'coerce': float}
-                }
-            }
-        }
+        'ceilings': Ceiling.SCHEMA,
+        'floors': Floor.SCHEMA
     }
 
     @classmethod
